프로필관련 리팩토링, 인증 목록조회, 취소
