+++ src/store/index.js
@@ -1,5 +1,5 @@
 import { createStore } from 'vuex';
-import axios from '../axios';
+import axios from '@/axios';
 
 export default createStore({
   state: {
